Added if statement to test install from TestPyPI
This if statement prevents a test install from TestPyPI if the PSR action did not create a new release. This is useful because it saves time, as well as prevents errors if no TestPyPI repository yet exists and PSR didn't make a version bump (can happen at the beginning of a project).